Hackathon_0/silver/AI_Employee_Vault/requirements.txt
# AI Employee - Silver Tier Requirements
# Python 3.13 or higher required

# Core Dependencies
watchdog>=4.0.0          # File system monitoring

# Gmail Integration (Optional)
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.115.0

# WhatsApp Integration (Optional - requires additional setup)
playwright>=1.41.0       # Browser automation for WhatsApp Web

# Utilities
python-dotenv>=1.0.0     # Environment variable management
requests>=2.31.0         # HTTP requests

# Development
pytest>=8.0.0            # Testing framework
